Move and refactor CreateDocumentSubmission tests to EvidenceRequests tests
